移植 1.21.1 Fabric

+++ src/main/java/org/mesdag/particlestorm/api/MolangInstance.java
@@ -18,13 +18,13 @@ public interface MolangInstance {
 
     float tickLifetime();
 
-    float getRandom1();
+    double getRandom1();
 
-    float getRandom2();
+    double getRandom2();
 
-    float getRandom3();
+    double getRandom3();
 
-    float getRandom4();
+    double getRandom4();
 
     ResourceLocation getIdentity();
